chore: remove market folder, move into markets
