+++ src/pages/Markets/Market.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from 'react';
 import styled from 'styled-components';
 import { RouteComponentProps } from 'react-router';
-import { TOKEN_LABELS } from '../../../constants/tokens';
+import { TOKEN_LABELS } from '../../constants/tokens';
 import { BuildInterface } from './Build';
-import { usePositionActionHandlers } from '../../../state/positions/hooks';
+import { usePositionActionHandlers } from '../../state/positions/hooks';
 
 export const Container = styled.div`
   display: flex;
